Fixed a loot of 'unreferenced formal parameter' and datatype conversion warnings

+++ drivers/can_tcp_win32/can_tcp_win32_server.cpp
@@ -54,7 +54,7 @@ unsigned __stdcall Connection(void* a) {
 
   delete s;
 
-  printf("Connection closed (0x%x).\n",s);
+  printf("Connection closed (0x%x).\n",(unsigned int)s);
 
   return 0;
 }
